Update records.py
Remove sections of the search functions that deals with optional parameters

Removes dummy data and creates a validation function for the passed ids to ensure they are numbers. 

Included try-catches for allow operations with the JSON object

Updated the parameters in the update_record function and altered the code to determine the record to update more accurately with only the changed information passed rather than every element in the json record.

Updated both the delete and update functions to accept partial parameters to account for flights. 

Implemented response codes for the GUI to determine the response to clients.